pyproject.toml
[project]
name = "envelope"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Jarek Liesen", email = "jarek.liesen@reuben.ox.ac.uk"},
]
dependencies = [
    "jax>=0.5.0",
]

license = {text = "MIT"}
keywords = [
    "machine-learning",
    "reinforcement-learning",
    "jax",
    "vectorization",
    "hardware-acceleration",
]

[project.urls]
Homepage = "https://github.com/keraJLi/envelope"
Repository = "https://github.com/keraJLi/envelope"
Documentation = "https://github.com/keraJLi/envelope#readme"
Issues = "https://github.com/keraJLi/envelope/issues"
Changelog = "https://github.com/keraJLi/envelope/releases"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
gymnax = { git = "https://github.com/RobertTLange/gymnax" }

[dependency-groups]
compat = [
    "gymnax @ git+https://github.com/RobertTLange/gymnax@main",
    "brax>=0.13.0",
    "craftax>=1.4.3",
    "navix>=0.7.0",
    "jumanji>=1.0.1",
    "kinetix-env>=2.0.0",
    "playground>=0.1.0",
]
dev = [
    "hypothesis>=6.148.1",
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.2",
]

[tool.pytest.ini_options]
markers = [
  "compat: tests requiring optional compat dependencies",
]
